reduce duplications and code smells
